데이터 직접삭제 -> 논리적 삭제로 변경

+++ src/main/java/com/example/demo/controller/ProductController.java
@@ -66,7 +66,7 @@ public class ProductController {
 
     @GetMapping("/product/{productNum}")
     public String getProductPage(@SessionAttribute(name = SessionConst.LOGIN_MEMBER,required = false) Member loginMember,
-                              @PathVariable int productNum, Model model){
+                              @PathVariable Long productNum, Model model){
         model.addAttribute("member",loginMember);
         Product product = productService.getProduct(productNum);
         model.addAttribute("product", product);
